.history/requirements_20220812214130.txt
streamlit
pandas
sklearn
sklean.feature_extraction